requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==8.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
